Introduced JWT authentication
Installed and configured JWT library to serve the purpose of issuing JWT, on successful user authentication.
Set up a service class called JwtIssuerService, which is responsible for handling all type of operations related to JWT. The class is registered as a service in Program.cs and is being injected in controllers through dependency injection.
The Signin method, part of the AuthController now returns an object, containing the JWT.